backend/requirements.txt
flask
flask-cors
python-dotenv
google-genai
sentence-transformers
faiss-cpu
pypdf
numpy